WebDriver BiDi: add tests for background in browsingContext.create
