+++ webdriver/tests/bidi/browsing_context/__init__.py
@@ -1,3 +1,7 @@
+from typing import Any, Mapping
+
+from webdriver.bidi.modules.script import ContextTarget
+
 from .. import (
     any_int,
     any_string,
@@ -64,3 +68,13 @@ def assert_navigation_info(event, expected_navigation_info):
 
     if "url" in expected_navigation_info:
         assert event["url"] == expected_navigation_info["url"]
+
+
+async def get_visibility_state(bidi_session, context: Mapping[str, Any]) -> str:
+    result = await bidi_session.script.call_function(
+        function_declaration="""() => {
+        return document.visibilityState;
+    }""",
+        target=ContextTarget(context["context"]),
+        await_promise=False)
+    return result["value"]
